+++ src/explorer/check_helpview.ts
@@ -1,12 +1,11 @@
 import { CancellationToken, Webview, WebviewView, WebviewViewProvider, WebviewViewResolveContext } from "vscode";
-import { CheckManager } from "../check_manager";
-import { TfSecCheck } from "../tfsec_check";
+import { CheckResult, CheckSeverity } from "./check_result";
 import { TfsecTreeItem } from "./tfsec_treeitem";
 
 export class TfsecHelpProvider implements WebviewViewProvider {
-    private view : Webview | undefined;
-    
-    resolveWebviewView(webviewView: WebviewView, context: WebviewViewResolveContext<unknown>, token: CancellationToken): void | Thenable<void> {
+    private view: Webview | undefined;
+
+    resolveWebviewView(webviewView: WebviewView, _context: WebviewViewResolveContext<unknown>, _token: CancellationToken): void | Thenable<void> {
         this.view = webviewView.webview;
         this.update(null);
     }
@@ -16,34 +15,50 @@ export class TfsecHelpProvider implements WebviewViewProvider {
             return;
         }
         if (item === null) {
-            return ;
+            return;
         }
-        const codeData = CheckManager.getInstance().get(item.code);
+        const codeData = item.check;
         if (codeData === undefined) {
             this.view.html = `
 <h2>No check data available</h2>
 This check may no longer be valid. Check your tfsec is the latest version.
-`
+`;
             return;
         }
 
+        if (item.contextValue === 'TFSEC_CODE') {
+            this.view.html = `
+            <h2>Select a specific instance for more details</h2>
+For more information about the issue found, select a specific instance.
+            `;
+            return;
+        }
         this.view.html = getHtml(codeData);
     }
 }
 
-function getHtml(codeData: TfSecCheck | undefined): string {
-    if (codeData === undefined) {
+function getHtml(codeData: CheckResult | CheckSeverity | undefined): string {
+    if (codeData === undefined || codeData instanceof CheckSeverity) {
         return "";
     }
-    return`
-    <h2>${codeData?.code}</h2>
+    return `
+    <h2>${codeData?.codeDescription}</h2>
     ${codeData?.summary}
+
+    <h3>ID</h3>
+    ${codeData?.code}
+
+    <h3>Severity</h3>
+    ${codeData?.severity}
 
     <h3>Impact</h3>
     ${codeData?.impact}
 
     <h3>Resolution</h3>
     ${codeData?.resolution}
+
+    <h3>Filename</h3>
+    ${codeData?.filename}
 
     <h3>More Information</h3>
     <a href="${codeData?.docUrl}">${codeData?.docUrl}</a>